kovan-sd: Use banked data outputs; send timestamps
This patch allows for more data to be read from the FPGA using fewer
pins.  It lets us read four banks of 16-bit registers.  Additionally,
we now store the timestamp, and also have multiple packet types.

It is the job of the CPU to decode the packet type and then to
properly format it for network transmission.

This also adds a "reset clock" pin, which can be used with SD reset to
synchronize the CPU clock to the FPGA clock.  Additionally, LCD_B[0] is
now used as a "clock overflow" flag; when the clock overflows, this line
will toggle, allowing the CPU to keep track beyond the 34-second epoch.