kovan-sd: Remove SD logger
There were timing issues related to the SD logger module, where it
would log multiple times.  If we want to log this data on the FPGA,
we should synchronize better with the SD protocol stream itself,
rather than trying to rely on the SD stream being uncorrupted (or
free from anomalies such as charge time.